Change double to long
`HX711::tare()` was calling `HX711::read_average()` and storing the result in a double. The return type for `HX711::read_average()`, however, was a long.

+++ HX711.cpp
@@ -97,7 +97,7 @@ float HX711::get_units(byte times) {
 }
 
 void HX711::tare(byte times) {
-	double sum = read_average(times);
+	long sum = read_average(times);
 	set_offset(sum);
 }
